add registration/authentication

+++ E-commerce-product-catalog/Models/Product.cs
@@ -2,7 +2,7 @@
 {
     public class Product
     {
-        public Product(Guid id, string name, string description, Guid categoryId, decimal price, decimal quantity)
+        public Product(Guid id, string name, string description, Guid categoryId, decimal price, int quantity)
         {
             Id = id;
             Name = name;
@@ -22,7 +22,7 @@
         public string Description { get; set; }
         public Guid CategoryId { get; set; }
         public decimal Price { get; set; }
-        public decimal Quantity { get; set; }
+        public int Quantity { get; set; }
         public Guid OwnerId { get; set; } = Guid.Empty;
